feat: Add react-datepicker npm dependency and update sales manager interface
The package.json file was updated to include the react-datepicker npm dependency with version 7.4.0. This will enable the use of a date picker component in the application. Additionally, the sales manager interface was modified to change the property names from birthDateD to birthDate and registratedSinceD to registratedSince. This improves the clarity and consistency of the interface.

Switchet to SSR on editing sales managers.